docs(skills): same-pitch overlap note; consolidate note-write-ordering e2e
- Skills: document that two same-pitch notes can't sound at once — Live
  truncates the earlier to end where the next starts (both kept), and same
  pitch+start collapses to one. Reflects the sort-before-write behavior.
- E2E: add an update-clip transform-only path case (a transform repositions a
  note out of start_time order; the re-add must stay sorted so a same-pitch
  onset overlap truncates instead of dropping a note), and consolidate both
  note-write-ordering e2e tests into one ppal-note-write-ordering.test.ts
  (keeps e2e/mcp/clip within the folder-size limit).
